Always use plain text in the input box
We don't want to mess with the style on the fly since it moves the
cursor around every time we change the content of the box
programmatically.  When we use html we can get some bold lines on
initial load, which is inconsistent with the behavior the rest of
the time.