Handle HTML character input better
Previously HTML characters (<, >, &) in the input caused bad behavior
because they got replaced by their &lt; etc. forms, which caused an
update of the input box and made the input ugly. We don't actually
need to escape the input box because it's plaintext now, so we should
undo those changes before checking if the input needs an update.